# Remove TeeWriter