## Description

This PR removes the `TeeWriter` method from the `Context` interface. The decision to remove this method is based on the assessment that it does not add significant value to the current implementation. The functionality provided `TeeWriter` can be effectively replaced and extended by middleware solutions. Specifically, the [foxdump](https://github.com/tigerwill90/foxdump) middleware is planned to implement similar functionality, offering a more modular and flexible approach to capturing and duplicating response bodies for logging, debugging, or any other purposes that require access to the response body data stream.

## Removed Method Detail

The `TeeWriter` method was designed to append an additional writer (sink) to which the response body would be simultaneously written, alongside the primary `ResponseWriter`. This method aimed to facilitate scenarios where capturing the response body for logging or analysis was necessary without disrupting the standard response flow to the client.

## Performance
````
goos: linux
goarch: amd64
pkg: github.com/tigerwill90/fox
cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz
                    │   old.txt    │               new.txt               │
                    │    sec/op    │    sec/op     vs base               │
StaticAll-16          11.16µ ±  1%   11.26µ ±  1%  +0.88% (p=0.003 n=10)
GithubParamsAll-16    83.33n ±  1%   82.90n ±  2%       ~ (p=0.165 n=10)
OverlappingRoute-16   98.27n ±  0%   94.97n ±  1%  -3.35% (p=0.000 n=10)
StaticParallel-16     10.29n ±  1%   10.29n ±  4%       ~ (p=0.926 n=10)
CatchAll-16           31.38n ±  1%   31.71n ±  1%  +1.04% (p=0.050 n=10)
CatchAllParallel-16   5.918n ± 12%   5.806n ± 22%       ~ (p=0.280 n=10)
CloneWith-16          70.51n ±  1%   69.78n ±  2%  -1.03% (p=0.028 n=10)
geomean               74.15n         73.62n        -0.71%
````